fix: add validating url in request object
